Fixed conv. warning due to integer promotion on smaller than int types

+++ src/libsac2c/global/globals.c
@@ -622,7 +622,7 @@ get_terminal_size (void)
         return 80;
 
     ioctl (STDERR_FILENO, TIOCGWINSZ, &ws);
-    return ws.ws_col > 4 ? ws.ws_col - 4 : 1;
+    return (unsigned short)(ws.ws_col > 4 ? ws.ws_col - 4 : 1);
 }
 
 /*
